requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pyserial>=3.5
requests>=2.31.0
geopy>=2.4.0
pynmea2>=1.19.0
numpy>=1.24.0
flask>=3.0.0
